requirements.txt
anyio==4.12.1
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.2
cloudscraper==1.2.71
colorama==0.4.6
contourpy==1.3.3
curl_cffi==0.14.0
cycler==0.12.1
distlib==0.4.0
filelock==3.25.0
Flask==3.1.3
fonttools==4.62.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
numpy==2.4.4
packaging==26.0
pandas==3.0.2
pillow==12.1.0
pipenv==2026.0.3
platformdirs==4.9.4
pycparser==3.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.0
six==1.17.0
threadpoolctl==3.6.0
tzdata==2026.1
urllib3==2.6.3
virtualenv==21.1.0
Werkzeug==3.1.8
pymongo
dnspython
gunicorn==21.2.0
python-dotenv